Add vim-airline back, uses patched font if available
This will check if the font is installed, if it's not it falls back to a
simpler appearance. If you're on terminal you need to make sure that
non-ascii is rendered with the patched font (if you have it installed)
otherwise vim will look aweful.

This supports OSX and maybe some installations of linux (assuming your
    font is installed at /usr/share/fonts or ~/.fonts). On windows or
other installations of linux it will gracefully fallback to the simpler
view.

[Closes #10]