Add tests for running tests
I ultimately want to add python/pytest support but felt that the code in
`lib/functions.vim` was too poorly factored to reasonably extend.

This is a start in the direction of adding tests to this repo and is in
no way finished. I want to push this for some early feedback.

Specifically it would be good to get feedback on:

  * Should `lib/functions.vim` be extended? Maybe it would be better to
  use [vim-test](https://github.com/janko-m/vim-test) and add vipe
  support to vim-test.
  * Is it worth while backfilling tests? Are there other areas of the
  project that would benefit from testing?
  * What is the best way to get these tests working on ci?